+++ R/registration_model.R
@@ -24,10 +24,9 @@
 #' head(registration_mod)
 #'
 registration_model <-
-    function(
-        sce_pseudo,
-        covars = NULL,
-        var_registration = "registration_variable") {
+    function(sce_pseudo,
+    covars = NULL,
+    var_registration = "registration_variable") {
         ## Specify a formula without an intercept
         if (is.null(covars)) {
             mat_formula <-